src/app/requirements.txt
gradio>=4.0.0
torch 
pillow 
transformers 
peft